pyproject.toml
[project]
name = "maid-runner"
version = "2.0.0"
description = "Tool-agnostic validation framework for the Manifest-driven AI Development (MAID) methodology. Validates that code artifacts align with declarative manifests."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Mamerto Fabian Jr.", email = "mamerto@codefrost.com"}
]
keywords = ["maid", "validation", "ai", "development", "manifest", "tdd", "architecture"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "jsonschema>=4.25.1",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/mamertofabian/maid-runner"
Repository = "https://github.com/mamertofabian/maid-runner"
Issues = "https://github.com/mamertofabian/maid-runner/issues"
Documentation = "https://github.com/mamertofabian/maid-runner#readme"

[project.optional-dependencies]
typescript = [
    "tree-sitter>=0.23.2",
    "tree-sitter-typescript>=0.23.2",
]
svelte = [
    "tree-sitter>=0.23.2",
    "tree-sitter-svelte>=1.0.2",
    "tree-sitter-typescript>=0.23.2",
]
all = [
    "tree-sitter>=0.23.2",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-svelte>=1.0.2",
]
watch = [
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "black>=25.1.0",
    "ruff>=0.13.0",
    "build>=1.0.0",
    "twine>=5.0.0",
    "watchdog>=6.0.0",
    "pre-commit>=4.0.0",
    "vulture>=2.0",
    "pytest-cov>=7.0.0",
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
    "tree-sitter>=0.23.2",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-svelte>=1.0.2",
]

[project.scripts]
maid = "maid_runner.cli.commands._main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["maid_runner*"]
exclude = ["tests*", "manifests*", "docs*", "examples*"]

[tool.setuptools.package-data]
maid_runner = ["py.typed", "schemas/*.json", "docs/maid_specs.md", "docs/unit-testing-rules.md", "claude/manifest.json", "claude/agents/*.md", "claude/commands/*.md"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]
